backend/requirements.txt
Django==3.2.6
djangorestframework==3.14.0
rest_framework_simplejwt==0.0.2
